+++ rbx_reflection/CHANGELOG.md
@@ -1,5 +1,8 @@
 # rbx_reflection Changelog
 
+## Unreleased Changes
+* Added `DataType::ty` to convert to `VariantType` infallibly.
+
 ## 6.0.0 (2025-10-10)
 * Updated to rbx_types 3.0.0. As this is a major version bump for a dependency, this crate has also gone up a major version to avoid mismatched versions. It is otherwise unchanged.